requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.3.0=pyhd8ed1ab_0
aiohttp=3.8.1=py310h5764c6d_1
aiosignal=1.3.1=pyhd8ed1ab_0
ale-py=0.8.0=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=py310h06a4308_0
attrs=22.1.0=pyh71513ae_1
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=pypi_0
black=22.6.0=py310h06a4308_0
blas=1.0=mkl
bleach=5.0.1=pypi_0
blinker=1.5=pyhd8ed1ab_0
box2d-py=2.3.5=pypi_0
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.9.24=ha878542_0
cachetools=5.2.0=pyhd8ed1ab_0
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.1=py310h74dc2b5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py310h06a4308_0
cloudpickle=2.2.0=pypi_0
contourpy=1.0.6=pypi_0
cryptography=38.0.1=py310h9ce1e76_0
cuda=11.6.2=0
cuda-cccl=11.6.55=hf6102b2_0
cuda-command-line-tools=11.6.2=0
cuda-compiler=11.6.2=0
cuda-cudart=11.6.55=he381448_0
cuda-cudart-dev=11.6.55=h42ad0f4_0
cuda-cuobjdump=11.6.124=h2eeebcb_0
cuda-cupti=11.6.124=h86345e5_0
cuda-cuxxfilt=11.6.124=hecbf4f6_0
cuda-driver-dev=11.6.55=0
cuda-gdb=11.8.86=0
cuda-libraries=11.6.2=0
cuda-libraries-dev=11.6.2=0
cuda-memcheck=11.8.86=0
cuda-nsight=11.8.86=0
cuda-nsight-compute=11.8.0=0
cuda-nvcc=11.6.124=hbba6d2d_0
cuda-nvdisasm=11.8.86=0
cuda-nvml-dev=11.6.55=haa9ef22_0
cuda-nvprof=11.8.87=0
cuda-nvprune=11.6.124=he22ec0a_0
cuda-nvrtc=11.6.124=h020bade_0
cuda-nvrtc-dev=11.6.124=h249d397_0
cuda-nvtx=11.6.124=h0630a44_0
cuda-nvvp=11.8.87=0
cuda-runtime=11.6.2=0
cuda-samples=11.6.101=h8efea70_0
cuda-sanitizer-api=11.8.86=0
cuda-toolkit=11.6.2=0
cuda-tools=11.6.2=0
cuda-visual-tools=11.6.2=0
cycler=0.11.0=pypi_0
cython=0.29.32=pypi_0
debugpy=1.5.1=py310h295c915_0
decorator=4.4.2=pypi_0
defusedxml=0.7.1=pypi_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.4=py310h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
fasteners=0.18=pypi_0
fastjsonschema=2.16.2=pypi_0
ffmpeg=4.2.2=h20bf706_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.2.0=py310h7f8727e_1
gds-tools=1.4.0.31=0
giflib=5.2.1=h7b6447c_0
gitdb=4.0.9=pypi_0
gitpython=3.1.29=pypi_0
glfw=2.5.5=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.14.0=pyh1a96a4e_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
grpcio=1.42.0=py310hce63b2e_0
gym=0.26.2=pypi_0
gym-notices=0.0.8=pypi_0
gymnasium=0.26.3=pypi_0
gymnasium-notices=0.0.1=pypi_0
htmlmin=0.1.12=pypi_0
idna=3.4=py310h06a4308_0
imagehash=4.3.1=pypi_0
imageio=2.22.3=pypi_0
imageio-ffmpeg=0.4.7=pypi_0
importlib-metadata=5.0.0=pyha770c72_1
importlib-resources=5.10.0=pypi_0
iniconfig=1.1.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.15.2=py310h06a4308_0
ipython=8.4.0=py310h06a4308_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.0.2=pypi_0
jedi=0.18.1=py310h06a4308_1
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jpeg=9e=h7f8727e_0
jsonschema=4.17.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-console=6.4.4=pypi_0
jupyter-server=1.21.0=pypi_0
jupyter_client=7.3.5=py310h06a4308_0
jupyter_core=4.11.1=py310h06a4308_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=3.0.3=pypi_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcufile=1.4.0.31=0
libcufile-dev=1.4.0.31=0
libcurand=10.3.0.86=0
libcurand-dev=10.3.0.86=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.8=h7f8727e_5
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libidn2=2.3.2=h7f8727e_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.20.1=h4ff587b_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.4.0=hecacb30_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libvpx=1.7.0=h439df22_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
lz4=4.0.2=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.4.1=pyhd8ed1ab_0
markupsafe=2.1.1=py310h5764c6d_1
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.6=py310h06a4308_0
missingno=0.5.1=pypi_0
mistune=2.0.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
moviepy=1.0.3=pypi_0
mujoco=2.2.0=pypi_0
mujoco-py=2.1.2.14=pypi_0
multidict=6.0.2=py310h5764c6d_1
multimethod=1.9=pypi_0
mypy_extensions=0.4.3=py310h06a4308_0
nbclassic=0.4.8=pypi_0
nbclient=0.7.0=pypi_0
nbconvert=7.2.3=pypi_0
nbformat=5.7.0=pypi_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py310h06a4308_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.8=pypi_0
notebook=6.5.2=pypi_0
notebook-shim=0.2.2=pypi_0
nsight-compute=2022.3.0.22=0
numpy=1.23.3=py310hd5efca6_1
numpy-base=1.23.3=py310h8e6c178_1
oauthlib=3.2.2=pyhd8ed1ab_0
opencv-python=4.6.0.66=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.5.1=pypi_0
pandas-profiling=3.4.0=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pyhd3eb1b0_0
pathspec=0.9.0=py310h06a4308_0
pathtools=0.1.2=pypi_0
patsy=0.5.3=pypi_0
pexpect=4.8.0=pyhd3eb1b0_3
phik=0.12.2=pypi_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py310hace64e9_1
pip=22.2.2=py310h06a4308_0
platformdirs=2.5.2=py310h06a4308_0
pluggy=1.0.0=pypi_0
proglog=0.1.10=pypi_0
prometheus-client=0.15.0=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=4.21.9=pypi_0
psutil=5.9.0=py310h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
py=1.11.0=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd3eb1b0_0
pygame=2.1.0=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.6.0=pyhd8ed1ab_0
pyopengl=3.1.6=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py310h06a4308_0
pyrsistent=0.19.2=pypi_0
pysocks=1.7.1=py310h06a4308_0
pytest=7.0.1=pypi_0
python=3.10.6=haa1d7c7_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.10=2_cp310
pytorch=1.13.0=py3.10_cuda11.6_cudnn8.3.2_0
pytorch-cuda=11.6=h867d48c_0
pytorch-mutex=1.0=cuda
pytz=2022.6=pypi_0
pyu2f=0.1.5=pyhd8ed1ab_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=23.2.0=py310h6a678d5_0
qtconsole=5.4.0=pypi_0
qtpy=2.2.1=pypi_0
readline=8.2=h5eee18b_0
requests=2.28.1=py310h06a4308_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
scipy=1.9.3=pypi_0
seaborn=0.12.1=pypi_0
send2trash=1.8.0=pypi_0
sentry-sdk=1.10.1=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=65.5.0=py310h06a4308_0
shortuuid=1.0.10=pypi_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.39.3=h5082296_0
stack_data=0.2.0=pyhd3eb1b0_0
statsmodels=0.13.5=pypi_0
swig=4.1.0=pypi_0
tangled-up-in-unicode=0.2.0=pypi_0
tensorboard=2.11.0=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py310h597c629_2
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
terminado=0.17.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=py310h06a4308_0
torch-tb-profiler=0.4.0=pypi_0
torchaudio=0.13.0=py310_cu116
torchvision=0.14.0=py310_cu116
tornado=6.2=py310h5eee18b_0
tqdm=4.64.1=pypi_0
traitlets=5.5.0=pypi_0
typing_extensions=4.3.0=py310h06a4308_0
tzdata=2022f=h04d1e81_0
urllib3=1.26.12=py310h06a4308_0
visions=0.7.5=pypi_0
wandb=0.13.5=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=pypi_0
websocket-client=1.4.2=pypi_0
werkzeug=2.2.2=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=4.0.3=pypi_0
x264=1!157.20191217=h7b6447c_0
xz=5.2.6=h5eee18b_0
yarl=1.7.2=py310h5764c6d_2
zeromq=4.3.4=h2531618_0
zipp=3.10.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
